Critical bug fix and major authorization updates for nbia.py
Fixed a formatting error that was preventing nbia.py from importing.  Also updated most nbia functions so that specifying an api_url is no longer necessary, and that NLST search results are automatically included in regular queries.  More updates are required for some functions, but wanted to get this published asap due to the import bug.